requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.4
cymem==2.0.8
datasets==2.16.1
deep-translator==1.11.4
dill==0.3.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
et-xmlfile==1.1.0
executing==2.0.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.5.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.60.0
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.2
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.11.0
imblearn==0.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
langcodes==3.3.0
libclang==16.0.6
lightning==2.1.2
lightning-utilities==0.9.0
lxml==5.1.0
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-tensorrt==99.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pathy==0.10.3
pillow==10.2.0
preshed==3.0.9
protobuf==4.23.4
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-docx==1.1.0
pytorch-lightning==2.1.2
pytorch-nlp==0.5.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.8
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.4.0
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics==1.2.0
torchvision==0.13.1+cu113
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unlimited-machine-translator==1.1.8
urllib3==2.1.0
wasabi==1.1.2
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.2
